// src/lib/supabase/types.d.ts
export type Database = {
// ...
  public: {
    Tables: {
// ...
      game_records: {
        Row: {
          black_player: string | null
          created_at: string
          end_time: string | null
          event: string | null
          game_hash: string
          handicap: string | null
          id: number
          move_count: number
          recorded_at: string
          result: string | null
          start_time: string | null
          user_id: string | null
          white_player: string | null
        }
        Insert: {
          black_player?: string | null
          created_at?: string
          end_time?: string | null
          event?: string | null
          game_hash: string
          handicap?: string | null
          id?: number
          move_count: number
          recorded_at?: string
          result?: string | null
          start_time?: string | null
          user_id?: string | null
          white_player?: string | null
        }
        Update: {
          black_player?: string | null
          created_at?: string
          end_time?: string | null
          event?: string | null
          game_hash?: string
          handicap?: string | null
          id?: number
          move_count?: number
          recorded_at?: string
          result?: string | null
          start_time?: string | null
          user_id?: string | null
          white_player?: string | null
        }
        Relationships: []
      }
      shogi_moves_statistics: {
        Row: {
          created_at: string
          id: number
          lose: boolean
          move: string
          sfenx: string
          timeout: boolean
          user_id: string | null
          win: boolean
        }
        Insert: {
          created_at?: string
          id?: number
          lose: boolean
          move?: string
          sfenx?: string
          timeout: boolean
          user_id?: string | null
          win: boolean
        }
        Update: {
          created_at?: string
          id?: number
          lose?: boolean
          move?: string
          sfenx?: string
          timeout?: boolean
          user_id?: string | null
          win?: boolean
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
// ...
}